buff shell company

+++ Assets/Scripts/ElectricTrap.cs
@@ -52,7 +52,7 @@ public class ElectricTrap : MonoBehaviour
 
                     if (playerData.equippedPatches.Contains(Patches.RENDING_BLOWS))
                     {
-                        enemy.GainDOT(emblemLibrary.rendingBlows.value);
+                        enemy.GainDOT((float)emblemLibrary.rendingBlows.value);
                     }
                 }
