+++ Assets/Scripts/EnemyBodyVFX.cs
@@ -68,7 +68,7 @@ public class EnemyBodyVFX : MonoBehaviour
 
     private void OnAttackImpact(object sender, System.EventArgs e)
     {
-        if (playerData.equippedPatches.Contains(Patches.SPELLSWORD) && playerData.mana > emblemLibrary.spellsword.value)
+        if (playerData.equippedPatches.Contains(Patches.SPELLSWORD) && playerData.mana > (float)emblemLibrary.spellsword.value)
         {
             magicImpact.Play();
         }
